+++ packages/core-data/src/entity-types/plugin.ts
@@ -28,11 +28,7 @@ declare module './base-entity-records' {
 			/**
 			 * The plugin author.
 			 */
-			author: ContextualField<
-				Record< string, string >,
-				'view' | 'edit',
-				C
-			>;
+			author: ContextualField< string, 'view' | 'edit', C >;
 			/**
 			 * Plugin author's website address.
 			 */