+++ packages/core-data/src/entity-types/post-status.ts
@@ -1,9 +1,8 @@
-import type { Context, OmitNevers } from './helpers';
+import type { Context, ContextualField, OmitNevers } from './helpers';
 import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
 
 declare module './base-entity-records' {
 	export namespace BaseEntityRecords {
-		/* eslint-disable-next-line @typescript-eslint/no-unused-vars */
 		export interface PostStatusObject< C extends Context > {
 			/**
 			 * The title for the status.
@@ -13,27 +12,27 @@ declare module './base-entity-records' {
 			/**
 			 * Whether posts with this status should be private.
 			 */
-			private: boolean;
+			private: ContextualField< boolean, 'edit', C >;
 
 			/**
 			 * Whether posts with this status should be protected.
 			 */
-			protected: boolean;
+			protected: ContextualField< boolean, 'edit', C >;
 
 			/**
 			 * Whether posts of this status should be shown in the front end of the site.
 			 */
-			public: boolean;
+			public: ContextualField< boolean, 'view' | 'edit', C >;
 
 			/**
 			 * Whether posts with this status should be publicly-queryable.
 			 */
-			queryable: boolean;
+			queryable: ContextualField< boolean, 'view' | 'edit', C >;
 
 			/**
 			 * Whether to include posts in the edit listing for their post type.
 			 */
-			show_in_list: boolean;
+			show_in_list: ContextualField< boolean, 'edit', C >;
 
 			/**
 			 * An alphanumeric identifier for the status.
@@ -43,7 +42,7 @@ declare module './base-entity-records' {
 			/**
 			 * Whether posts of this status may have floating published dates.
 			 */
-			date_floating: boolean;
+			date_floating: ContextualField< boolean, 'view' | 'edit', C >;
 		}
 	}
 }